Reorganize settings page structure
- Rename "User Settings" to "User Preferences"
- Create new Navigation tab with Navigation Menu and Homepage Carousels
- Move Measurement Units into View Preferences section
- Reorder Server Settings: Server Configuration now comes before User Management
